+++ R/plot_cusum.R
@@ -1,6 +1,6 @@
 #' Plot CUSUM chart for a cusum Object
 #'
-#' Produces a CUSUM chart. 
+#' Produces a CUSUM chart.
 #' @export
 #' @import utils
 #' @param x An object of class cusum
@@ -10,30 +10,48 @@
 
 
 plot.cusum <- function(x, signal = TRUE, ...) {
-  
-  plot(x = x$t, 
-       y = x$ct, 
-       type = "n",
-       xlim = c(0,max(x$t)),
-       ylim = c(0, max(x$ct)), 
-       ylab = expression(CUSUM[t]), xlab = "t",
-       ...)
-    
-  if (signal == TRUE){
-    points(x = x$t[x$signal == 1], 
-           y = x$ct[x$signal == 1], 
-           col = "red", 
-           cex = 2, 
-           pch = 8,
-           lwd = 5)
+  if (mean(x$ct) < 0){
+    y_lim <- c(min(x$ct), 0)
+  } else {
+    y_lim <- c(0, max(x$ct))
+  }
+
+  plot(
+    x = x$t,
+    y = x$ct,
+    type = "n",
+    xlim = c(0, max(x$t)),
+    ylim = y_lim,
+    ylab = expression(CUSUM[t]), xlab = "t",
+    ...
+  )
+
+  if (signal == TRUE) {
+    points(
+      x = x$t[x$signal == 1],
+      y = x$ct[x$signal == 1],
+      col = "red",
+      cex = 2,
+      pch = 8,
+      lwd = 5
+    )
   }
   
-  lines(x = x$t, 
-        y = x$ct)
+  lines(
+    x = x$t,
+    y = x$ct
+  )
+
+  lines(
+    x = x$t,
+    y = x$limit,
+    col = "Blue"
+  )
   
-  abline(h = unique(x$limit), col = "Blue")
-  points(x = x$t, 
-         y = x$ct, 
-         cex = .5,
-         pch = 16)
+  points(
+    x = x$t,
+    y = x$ct,
+    cex = .5,
+    pch = 16
+  )
 }