+++ NEWS.md
@@ -1,3 +1,10 @@
+# cusum 0.3.0 (2019/06/17)
+
+### New features
+* new function to calculate CUSUM control limits exactly for very small sample sizes
+* function to calculate DPCL following Zhang and Woodall (2016) for RA-CUSUM
+* detect process improvements with cusum() and racusum()
+
 # cusum 0.2.1 (2019/03/19)
 
 ### Bug fixes